+++ x/ugov/events.pb.go
@@ -5,6 +5,7 @@ package ugov
 
 import (
 	fmt "fmt"
+	_ "github.com/cosmos/cosmos-proto"
 	types "github.com/cosmos/cosmos-sdk/types"
 	proto "github.com/cosmos/gogoproto/proto"
 	_ "github.com/gogo/protobuf/gogoproto"
@@ -24,23 +25,23 @@ var _ = math.Inf
 // proto package needs to be updated.
 const _ = proto.GoGoProtoPackageIsVersion3 // please upgrade the proto package
 
-// EventMinTxFees is emitted when MsgGovUpdateMinFees is correctly executed.
-type EventMinTxFees struct {
-	MinTxFees []types.DecCoin `protobuf:"bytes,1,rep,name=min_tx_fees,json=minTxFees,proto3" json:"min_tx_fees"`
+// EventMinGasPrice is emitted when MsgGovUpdateMinGasPrice is correctly executed.
+type EventMinGasPrice struct {
+	MinGasPrices []types.DecCoin `protobuf:"bytes,1,rep,name=min_gas_prices,json=minGasPrices,proto3" json:"min_gas_prices"`
 }
 
-func (m *EventMinTxFees) Reset()         { *m = EventMinTxFees{} }
-func (m *EventMinTxFees) String() string { return proto.CompactTextString(m) }
-func (*EventMinTxFees) ProtoMessage()    {}
-func (*EventMinTxFees) Descriptor() ([]byte, []int) {
+func (m *EventMinGasPrice) Reset()         { *m = EventMinGasPrice{} }
+func (m *EventMinGasPrice) String() string { return proto.CompactTextString(m) }
+func (*EventMinGasPrice) ProtoMessage()    {}
+func (*EventMinGasPrice) Descriptor() ([]byte, []int) {
 	return fileDescriptor_0885cdf0808da4ea, []int{0}
 }
-func (m *EventMinTxFees) XXX_Unmarshal(b []byte) error {
+func (m *EventMinGasPrice) XXX_Unmarshal(b []byte) error {
 	return m.Unmarshal(b)
 }
-func (m *EventMinTxFees) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+func (m *EventMinGasPrice) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
 	if deterministic {
-		return xxx_messageInfo_EventMinTxFees.Marshal(b, m, deterministic)
+		return xxx_messageInfo_EventMinGasPrice.Marshal(b, m, deterministic)
 	} else {
 		b = b[:cap(b)]
 		n, err := m.MarshalToSizedBuffer(b)
@@ -50,44 +51,88 @@ func (m *EventMinTxFees) XXX_Marshal(b []byte, deterministic bool) ([]byte, erro
 		return b[:n], nil
 	}
 }
-func (m *EventMinTxFees) XXX_Merge(src proto.Message) {
-	xxx_messageInfo_EventMinTxFees.Merge(m, src)
+func (m *EventMinGasPrice) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_EventMinGasPrice.Merge(m, src)
 }
-func (m *EventMinTxFees) XXX_Size() int {
+func (m *EventMinGasPrice) XXX_Size() int {
 	return m.Size()
 }
-func (m *EventMinTxFees) XXX_DiscardUnknown() {
-	xxx_messageInfo_EventMinTxFees.DiscardUnknown(m)
+func (m *EventMinGasPrice) XXX_DiscardUnknown() {
+	xxx_messageInfo_EventMinGasPrice.DiscardUnknown(m)
 }
 
-var xxx_messageInfo_EventMinTxFees proto.InternalMessageInfo
+var xxx_messageInfo_EventMinGasPrice proto.InternalMessageInfo
+
+// EventEmergencyGroup is emitted when MsgGovSetEmergencyGroup is correctly executed.
+type EventEmergencyGroup struct {
+	EmergencyGroup string `protobuf:"bytes,1,opt,name=emergency_group,json=emergencyGroup,proto3" json:"emergency_group,omitempty"`
+}
+
+func (m *EventEmergencyGroup) Reset()         { *m = EventEmergencyGroup{} }
+func (m *EventEmergencyGroup) String() string { return proto.CompactTextString(m) }
+func (*EventEmergencyGroup) ProtoMessage()    {}
+func (*EventEmergencyGroup) Descriptor() ([]byte, []int) {
+	return fileDescriptor_0885cdf0808da4ea, []int{1}
+}
+func (m *EventEmergencyGroup) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *EventEmergencyGroup) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_EventEmergencyGroup.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *EventEmergencyGroup) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_EventEmergencyGroup.Merge(m, src)
+}
+func (m *EventEmergencyGroup) XXX_Size() int {
+	return m.Size()
+}
+func (m *EventEmergencyGroup) XXX_DiscardUnknown() {
+	xxx_messageInfo_EventEmergencyGroup.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_EventEmergencyGroup proto.InternalMessageInfo
 
 func init() {
-	proto.RegisterType((*EventMinTxFees)(nil), "umee.ugov.v1.EventMinTxFees")
+	proto.RegisterType((*EventMinGasPrice)(nil), "umee.ugov.v1.EventMinGasPrice")
+	proto.RegisterType((*EventEmergencyGroup)(nil), "umee.ugov.v1.EventEmergencyGroup")
 }
 
 func init() { proto.RegisterFile("umee/ugov/v1/events.proto", fileDescriptor_0885cdf0808da4ea) }
 
 var fileDescriptor_0885cdf0808da4ea = []byte{
-	// 236 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x34, 0x8f, 0xb1, 0x4a, 0xc4, 0x40,
-	0x10, 0x40, 0x37, 0x28, 0x82, 0x39, 0xb1, 0x38, 0x2c, 0xf4, 0x90, 0x51, 0x2c, 0xc4, 0xc6, 0x19,
-	0xa2, 0xf8, 0x03, 0xf1, 0xb4, 0xb3, 0x91, 0xab, 0x6c, 0x8e, 0x4b, 0x18, 0xe3, 0x22, 0xbb, 0x23,
-	0xb7, 0x9b, 0x35, 0x9f, 0xe1, 0x67, 0xa5, 0xbc, 0xd2, 0x4a, 0x34, 0xf9, 0x11, 0xd9, 0x44, 0xbb,
-	0x19, 0xde, 0xf0, 0x98, 0x97, 0x1e, 0xd5, 0x86, 0x99, 0xea, 0x4a, 0x02, 0x85, 0x8c, 0x38, 0xb0,
-	0xf5, 0x0e, 0xdf, 0xd6, 0xe2, 0x65, 0xba, 0x17, 0x11, 0x46, 0x84, 0x21, 0x9b, 0x1d, 0x54, 0x52,
-	0xc9, 0x00, 0x28, 0x4e, 0xe3, 0xcd, 0x0c, 0x4a, 0x71, 0x46, 0x1c, 0x15, 0x2b, 0xc7, 0x14, 0xb2,
-	0x82, 0xfd, 0x2a, 0xa3, 0x52, 0xb4, 0x1d, 0xf9, 0xd9, 0x22, 0xdd, 0xbf, 0x8b, 0xce, 0x07, 0x6d,
-	0x17, 0xcd, 0x3d, 0xb3, 0x9b, 0xe6, 0xe9, 0xc4, 0x68, 0xbb, 0xf4, 0xcd, 0xf2, 0x99, 0xd9, 0x1d,
-	0x26, 0xa7, 0x5b, 0x17, 0x93, 0xab, 0x63, 0x1c, 0x3d, 0x18, 0x3d, 0xf8, 0xe7, 0xc1, 0x39, 0x97,
-	0xb7, 0xa2, 0x6d, 0xbe, 0xdd, 0x7e, 0x9d, 0xa8, 0xc7, 0x5d, 0xf3, 0xef, 0xc8, 0xe7, 0xed, 0x0f,
-	0xa8, 0xb6, 0x83, 0x64, 0xd3, 0x41, 0xf2, 0xdd, 0x41, 0xf2, 0xd1, 0x83, 0xda, 0xf4, 0xa0, 0x3e,
-	0x7b, 0x50, 0x4f, 0xe7, 0x95, 0xf6, 0x2f, 0x75, 0x81, 0xa5, 0x18, 0x8a, 0x09, 0x97, 0x96, 0xfd,
-	0xbb, 0xac, 0x5f, 0x87, 0x85, 0xc2, 0x0d, 0x35, 0x43, 0x6f, 0xb1, 0x33, 0xbc, 0x78, 0xfd, 0x1b,
-	0x00, 0x00, 0xff, 0xff, 0x49, 0x9a, 0xdb, 0x28, 0x03, 0x01, 0x00, 0x00,
+	// 308 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x54, 0x90, 0x41, 0x4a, 0xc3, 0x40,
+	0x18, 0x85, 0x13, 0x14, 0xc1, 0x58, 0xaa, 0xd4, 0x2e, 0xda, 0x22, 0x63, 0xe9, 0x42, 0xba, 0xe9,
+	0x0c, 0x51, 0x3c, 0x40, 0x6b, 0x4b, 0xdd, 0x08, 0x52, 0x37, 0x22, 0x42, 0x49, 0xa6, 0x3f, 0xe3,
+	0x20, 0x99, 0xbf, 0xcc, 0x24, 0xa3, 0xde, 0xc2, 0xc3, 0x78, 0x88, 0x2e, 0x8b, 0x2b, 0x57, 0xa2,
+	0xcd, 0x45, 0x64, 0x92, 0x88, 0xb8, 0xfb, 0xdf, 0x7b, 0x3f, 0xdf, 0x83, 0x17, 0xb4, 0xb3, 0x04,
+	0x80, 0x65, 0x02, 0x2d, 0xb3, 0x21, 0x03, 0x0b, 0x2a, 0x35, 0x74, 0xa9, 0x31, 0xc5, 0x46, 0xcd,
+	0x45, 0xd4, 0x45, 0xd4, 0x86, 0x1d, 0xc2, 0xd1, 0x24, 0x68, 0x58, 0x1c, 0x19, 0x60, 0x36, 0x8c,
+	0x21, 0x8d, 0x42, 0xc6, 0x51, 0xaa, 0xf2, 0xbb, 0xd3, 0x2e, 0xf3, 0x79, 0xa1, 0x58, 0x29, 0xaa,
+	0xa8, 0x29, 0x50, 0x60, 0xe9, 0xbb, 0xab, 0x74, 0x7b, 0xf7, 0xc1, 0xc1, 0xc4, 0xd5, 0x5d, 0x49,
+	0x35, 0x8d, 0xcc, 0xb5, 0x96, 0x1c, 0x1a, 0x97, 0x41, 0x3d, 0x91, 0x6a, 0x2e, 0x22, 0xc7, 0x91,
+	0x1c, 0x4c, 0xcb, 0xef, 0x6e, 0xf5, 0xf7, 0x4e, 0x8f, 0x68, 0x05, 0x74, 0xed, 0xb4, 0x6a, 0xa7,
+	0x63, 0xe0, 0x17, 0x28, 0xd5, 0x68, 0x7b, 0xf5, 0x79, 0xec, 0xcd, 0x6a, 0xc9, 0x1f, 0xc8, 0xf4,
+	0x6e, 0x83, 0xc3, 0x82, 0x3e, 0x49, 0x40, 0x0b, 0x50, 0xfc, 0x65, 0xaa, 0x31, 0x5b, 0x36, 0x86,
+	0xc1, 0x3e, 0xfc, 0x3a, 0x73, 0xe1, 0xac, 0x96, 0xdf, 0xf5, 0xfb, 0xbb, 0xa3, 0xd6, 0xfb, 0xdb,
+	0xa0, 0x59, 0x95, 0x0c, 0x17, 0x0b, 0x0d, 0xc6, 0xdc, 0xa4, 0x5a, 0x2a, 0x31, 0xab, 0xc3, 0x3f,
+	0xc4, 0x68, 0xbc, 0xfa, 0x26, 0xde, 0x6a, 0x43, 0xfc, 0xf5, 0x86, 0xf8, 0x5f, 0x1b, 0xe2, 0xbf,
+	0xe6, 0xc4, 0x5b, 0xe7, 0xc4, 0xfb, 0xc8, 0x89, 0x77, 0x77, 0x22, 0x64, 0xfa, 0x90, 0xc5, 0x94,
+	0x63, 0xc2, 0xdc, 0x7e, 0x03, 0x05, 0xe9, 0x13, 0xea, 0xc7, 0x42, 0x30, 0x7b, 0xce, 0x9e, 0x8b,
+	0xb1, 0xe3, 0x9d, 0x62, 0x84, 0xb3, 0x9f, 0x00, 0x00, 0x00, 0xff, 0xff, 0xa2, 0x47, 0x3b, 0x1f,
+	0x80, 0x01, 0x00, 0x00,
 }
 
-func (m *EventMinTxFees) Marshal() (dAtA []byte, err error) {
+func (m *EventMinGasPrice) Marshal() (dAtA []byte, err error) {
 	size := m.Size()
 	dAtA = make([]byte, size)
 	n, err := m.MarshalToSizedBuffer(dAtA[:size])
@@ -97,20 +142,20 @@ func (m *EventMinTxFees) Marshal() (dAtA []byte, err error) {
 	return dAtA[:n], nil
 }
 
-func (m *EventMinTxFees) MarshalTo(dAtA []byte) (int, error) {
+func (m *EventMinGasPrice) MarshalTo(dAtA []byte) (int, error) {
 	size := m.Size()
 	return m.MarshalToSizedBuffer(dAtA[:size])
 }
 
-func (m *EventMinTxFees) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+func (m *EventMinGasPrice) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	i := len(dAtA)
 	_ = i
 	var l int
 	_ = l
-	if len(m.MinTxFees) > 0 {
-		for iNdEx := len(m.MinTxFees) - 1; iNdEx >= 0; iNdEx-- {
+	if len(m.MinGasPrices) > 0 {
+		for iNdEx := len(m.MinGasPrices) - 1; iNdEx >= 0; iNdEx-- {
 			{
-				size, err := m.MinTxFees[iNdEx].MarshalToSizedBuffer(dAtA[:i])
+				size, err := m.MinGasPrices[iNdEx].MarshalToSizedBuffer(dAtA[:i])
 				if err != nil {
 					return 0, err
 				}
@@ -120,6 +165,36 @@ func (m *EventMinTxFees) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 			i--
 			dAtA[i] = 0xa
 		}
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *EventEmergencyGroup) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *EventEmergencyGroup) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *EventEmergencyGroup) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.EmergencyGroup) > 0 {
+		i -= len(m.EmergencyGroup)
+		copy(dAtA[i:], m.EmergencyGroup)
+		i = encodeVarintEvents(dAtA, i, uint64(len(m.EmergencyGroup)))
+		i--
+		dAtA[i] = 0xa
 	}
 	return len(dAtA) - i, nil
 }
@@ -135,17 +210,30 @@ func encodeVarintEvents(dAtA []byte, offset int, v uint64) int {
 	dAtA[offset] = uint8(v)
 	return base
 }
-func (m *EventMinTxFees) Size() (n int) {
+func (m *EventMinGasPrice) Size() (n int) {
 	if m == nil {
 		return 0
 	}
 	var l int
 	_ = l
-	if len(m.MinTxFees) > 0 {
-		for _, e := range m.MinTxFees {
+	if len(m.MinGasPrices) > 0 {
+		for _, e := range m.MinGasPrices {
 			l = e.Size()
 			n += 1 + l + sovEvents(uint64(l))
 		}
+	}
+	return n
+}
+
+func (m *EventEmergencyGroup) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.EmergencyGroup)
+	if l > 0 {
+		n += 1 + l + sovEvents(uint64(l))
 	}
 	return n
 }
@@ -156,7 +244,7 @@ func sovEvents(x uint64) (n int) {
 func sozEvents(x uint64) (n int) {
 	return sovEvents(uint64((x << 1) ^ uint64((int64(x) >> 63))))
 }
-func (m *EventMinTxFees) Unmarshal(dAtA []byte) error {
+func (m *EventMinGasPrice) Unmarshal(dAtA []byte) error {
 	l := len(dAtA)
 	iNdEx := 0
 	for iNdEx < l {
@@ -179,15 +267,15 @@ func (m *EventMinTxFees) Unmarshal(dAtA []byte) error {
 		fieldNum := int32(wire >> 3)
 		wireType := int(wire & 0x7)
 		if wireType == 4 {
-			return fmt.Errorf("proto: EventMinTxFees: wiretype end group for non-group")
+			return fmt.Errorf("proto: EventMinGasPrice: wiretype end group for non-group")
 		}
 		if fieldNum <= 0 {
-			return fmt.Errorf("proto: EventMinTxFees: illegal tag %d (wire type %d)", fieldNum, wire)
+			return fmt.Errorf("proto: EventMinGasPrice: illegal tag %d (wire type %d)", fieldNum, wire)
 		}
 		switch fieldNum {
 		case 1:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field MinTxFees", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field MinGasPrices", wireType)
 			}
 			var msglen int
 			for shift := uint(0); ; shift += 7 {
@@ -214,10 +302,92 @@ func (m *EventMinTxFees) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.MinTxFees = append(m.MinTxFees, types.DecCoin{})
-			if err := m.MinTxFees[len(m.MinTxFees)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
+			m.MinGasPrices = append(m.MinGasPrices, types.DecCoin{})
+			if err := m.MinGasPrices[len(m.MinGasPrices)-1].Unmarshal(dAtA[iNdEx:postIndex]); err != nil {
 				return err
 			}
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipEvents(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *EventEmergencyGroup) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowEvents
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: EventEmergencyGroup: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: EventEmergencyGroup: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field EmergencyGroup", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvents
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvents
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvents
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.EmergencyGroup = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex